功能追加 1.为model统一添加equals和hashCode 2.将model.station.stationNameCN更名为stationName，将DTO.stationDTO.stationNameCN更名为stationName 3.完善user相关，现在可以验证用户和密码 4.增加StationInfoDTO类 5.更新了前端内容：基础的网页跳转和后端互动 6.引入Sa-token进行鉴权 7.创建了permission数据库表及其对应的javaBean和mapper

+++ src/main/java/com/RailManager/demo/controller/LineController.java
@@ -1,12 +1,15 @@
 package com.RailManager.demo.controller;
 
+import cn.dev33.satoken.stp.StpUtil;
 import com.RailManager.demo.DTO.LineDTO;
 import com.RailManager.demo.DTO.LineInfoDTO;
 import com.RailManager.demo.DTO.StationDTO;
+import com.RailManager.demo.DTO.StationInfoDTO;
 import com.RailManager.demo.model.Line;
 import com.RailManager.demo.model.Station;
 import com.RailManager.demo.service.RailwayService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,19 +24,35 @@ public class LineController {
     @Autowired
     RailwayService railwayService;
 
-    final private ResponseEntity<String> crossDomain = ResponseEntity.ok().header("Access-Control-Allow-Origin" ,"localhost:8080").body("");
+    @Value("${crossdomain}")
+    private String crossDomainValues;
 
     @RequestMapping("/line/{lineName}")
     public ResponseEntity<LineInfoDTO> getLine(@PathVariable("lineName") String lineName){
         return ResponseEntity.ok()
-                .headers(crossDomain.getHeaders())
+                .header("Access-Control-Allow-Origin", crossDomainValues)
                 .body(railwayService.getLineInfoDTOByName(lineName));
     }
-    @RequestMapping("/line")
-    public ResponseEntity<List<LineInfoDTO>> getLine(){
+    @RequestMapping("/line/{lineName}/{stationName}")
+    public ResponseEntity<StationInfoDTO> getStationByNameAndLine(
+            @PathVariable("lineName") String lineName,
+            @PathVariable("stationName")String stationName){
         return ResponseEntity.ok()
-                .headers(crossDomain.getHeaders())
+                .header("Access-Control-Allow-Origin", crossDomainValues)
+                .body(railwayService.getStationInfoDTOByNameAndLine(stationName, lineName));
+    }
+
+    @RequestMapping("/line")
+    public ResponseEntity<List<LineInfoDTO>> getLines(){
+        return ResponseEntity.ok()
+                .header("Access-Control-Allow-Origin", crossDomainValues)
                 .body(railwayService.getAllLineInfoDTO());
+    }
+    @RequestMapping("/line/station")
+    public ResponseEntity<List<StationInfoDTO>> getStations(){
+        return ResponseEntity.ok()
+                .header("Access-Control-Allow-Origin", crossDomainValues)
+                .body(railwayService.getAllStationInfoDTO());
     }
 
     @RequestMapping("/line/download")
@@ -41,35 +60,43 @@ public class LineController {
         byte[] bytes = railwayService.getAllLineInfoDTO().toString().getBytes(StandardCharsets.UTF_8);
         return ResponseEntity.ok()
                 .header("Content-Disposition", "attachment; filename=XLM_lines_map.txt")
-                .headers(crossDomain.getHeaders())
+                .header("Access-Control-Allow-Origin", crossDomainValues)
                 .contentType(MediaType.APPLICATION_OCTET_STREAM)
                 .contentLength(bytes.length)
                 .body(bytes);
     }
 
-    @RequestMapping("/addLine")
+    @RequestMapping("/line/addLine")
     public ResponseEntity<Line> addLine(LineDTO dto){
-            return ResponseEntity.ok()
-                    .headers(crossDomain.getHeaders())
-                    .body(railwayService.addLine(dto));
-    }
-    @RequestMapping("/addStation")
-    public ResponseEntity<Station> addStation(StationDTO dto){
+        if(!StpUtil.hasPermission("canAdd"))
+            return ResponseEntity.ok().header("Access-Control-Allow-Origin", crossDomainValues).body(null);
         return ResponseEntity.ok()
-                .headers(crossDomain.getHeaders())
+                .header("Access-Control-Allow-Origin", crossDomainValues)
+                .body(railwayService.addLine(dto));
+    }
+    @RequestMapping("/line/addStation")
+    public ResponseEntity<Station> addStation(StationDTO dto){
+        if(!StpUtil.hasPermission("canAdd"))
+            return ResponseEntity.ok().header("Access-Control-Allow-Origin", crossDomainValues).body(null);
+        return ResponseEntity.ok()
+                .header("Access-Control-Allow-Origin", crossDomainValues)
                 .body(railwayService.addStation(dto));
     }
 
-    @RequestMapping("/delLine")
+    @RequestMapping("/line/delLine")
     public ResponseEntity<LineInfoDTO> delLine(String lineName){
+        if(!StpUtil.hasPermission("canDelete"))
+            return ResponseEntity.ok().header("Access-Control-Allow-Origin", crossDomainValues).body(null);
         return ResponseEntity.ok()
-                .headers(crossDomain.getHeaders())
+                .header("Access-Control-Allow-Origin", crossDomainValues)
                 .body(railwayService.deleteLine(lineName));
     }
-    @RequestMapping("/delStation")
+    @RequestMapping("/line/delStation")
     public ResponseEntity<Station> delStation(Integer stationId){
+        if(!StpUtil.hasPermission("canDelete"))
+            return ResponseEntity.ok().header("Access-Control-Allow-Origin", crossDomainValues).body(null);
         return ResponseEntity.ok()
-                .headers(crossDomain.getHeaders())
+                .header("Access-Control-Allow-Origin", crossDomainValues)
                 .body(railwayService.deleteStation(stationId));
     }
 }
